перемещение переменных из api.js в файл index.js
