+++ components/add-post-page-component.js
@@ -2,7 +2,7 @@ import { renderHeaderComponent } from './header-component.js'
 import { renderUploadImageComponent } from './upload-image-component.js'
 
 export function renderAddPostPageComponent({ appEl, onAddPostClick }) {
-    // let imageUrl = ''
+    let imageUrl = ''
 
     const render = () => {
         // @TODO: Реализовать страницу добавления поста
