+++ index.js
@@ -20,6 +20,17 @@ export let user = getUserFromLocalStorage()
 export let page = null
 export let posts = []
 
+export let isToken = ''
+export let userName = ''
+
+export const updateToken = (newToken) => {
+    isToken = newToken
+}
+
+export const updateUserName = (newName) => {
+    userName = newName
+}
+
 const getToken = () => {
     const token = user ? `Bearer ${user.token}` : undefined
     return token
